#967 container widgets reworked
- moved internals of container widget to corresponding folder
- moved internals of dynamic table to corresponding folder
