+++ ng2-components/ng2-activiti-form/src/components/widgets/container/container-column.model.ts
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-import { FormFieldModel } from './form-field.model';
+import { FormFieldModel } from './../core/form-field.model';
 
 export class ContainerColumnModel {
