+++ ng2-components/ng2-activiti-form/src/components/widgets/core/container.model.spec.ts
@@ -17,7 +17,6 @@
 
 import { ContainerModel } from './container.model';
 import { FormModel } from './form.model';
-import { FormFieldTypes } from './form-field-types';
 
 describe('ContainerModel', () => {
 
@@ -31,115 +30,6 @@ describe('ContainerModel', () => {
         let json = {};
         let model = new ContainerModel(null, json);
         expect(model.json).toBe(json);
-    });
-
-    it('should have 1 column layout by default', () => {
-        let container = new ContainerModel(null, null);
-        expect(container.numberOfColumns).toBe(1);
-    });
-
-    it('should be expanded by default', () => {
-        let container = new ContainerModel(null, null);
-        expect(container.isExpanded).toBeTruthy();
-    });
-
-    it('should setup with json config', () => {
-        let json = {
-            fieldType: '<type>',
-            id: '<id>',
-            name: '<name>',
-            type: '<type>',
-            tab: '<tab>',
-            numberOfColumns: 2,
-            params: {}
-        };
-        let container = new ContainerModel(null, json);
-        Object.keys(json).forEach(key => {
-            expect(container[key]).toEqual(json[key]);
-        });
-    });
-
-    it('should wrap fields into columns on setup', () => {
-        let form = new FormModel();
-        let json = {
-            fieldType: '<type>',
-            id: '<id>',
-            name: '<name>',
-            type: '<type>',
-            tab: '<tab>',
-            numberOfColumns: 3,
-            params: {},
-            visibilityCondition: {},
-            fields: {
-                '1': [
-                    { id: 'field-1' },
-                    { id: 'field-3' }
-                ],
-                '2': [
-                    { id: 'field-2' }
-                ],
-                '3': null
-            }
-        };
-        let container = new ContainerModel(form, json);
-        expect(container.columns.length).toBe(3);
-
-        let col1 = container.columns[0];
-        expect(col1.fields.length).toBe(2);
-        expect(col1.fields[0].id).toBe('field-1');
-        expect(col1.fields[1].id).toBe('field-3');
-
-        let col2 = container.columns[1];
-        expect(col2.fields.length).toBe(1);
-        expect(col2.fields[0].id).toBe('field-2');
-
-        let col3 = container.columns[2];
-        expect(col3.fields.length).toBe(0);
-    });
-
-    it('should allow collapsing only when of a group type', () => {
-        let container = new ContainerModel(new FormModel(), {
-            type:  FormFieldTypes.CONTAINER,
-            params: {
-                allowCollapse: true
-            }
-        });
-
-        expect(container.isCollapsible()).toBeFalsy();
-        container = new ContainerModel(new FormModel(), {
-            type:  FormFieldTypes.GROUP,
-            params: {
-                allowCollapse: true
-            }
-        });
-        expect(container.isCollapsible()).toBeTruthy();
-    });
-
-    it('should allow collapsing only when explicitly defined in params', () => {
-        let container = new ContainerModel(new FormModel(), {
-            type:  FormFieldTypes.GROUP,
-            params: {}
-        });
-        expect(container.isCollapsible()).toBeFalsy();
-
-        container = new ContainerModel(new FormModel(), {
-            type:  FormFieldTypes.GROUP,
-            params: {
-                allowCollapse: true
-            }
-        });
-        expect(container.isCollapsible()).toBeTruthy();
-    });
-
-    it('should be collapsed by default', () => {
-        let container = new ContainerModel(new FormModel(), {
-            type:  FormFieldTypes.GROUP,
-            params: {
-                allowCollapse: true,
-                collapseByDefault: true
-            }
-        });
-        expect(container.isCollapsedByDefault()).toBeTruthy();
     });
 
 });
