+++ ng2-components/ng2-activiti-form/src/components/widgets/core/form-field.model.ts
@@ -37,6 +37,7 @@ import {
 
 declare var moment: any;
 
+// Maps to FormFieldRepresentation
 export class FormFieldModel extends FormWidgetModel {
 
     private _value: string;
