+++ ng2-components/ng2-activiti-form/src/components/widgets/core/index.ts
@@ -22,12 +22,8 @@ export * from './form-field-option';
 export * from './form-widget.model';
 export * from './form-field.model';
 export * from './form.model';
-export * from './container-column.model';
 export * from './container.model';
 export * from './tab.model';
 export * from './form-outcome.model';
 export * from './form-outcome-event.model';
 export * from './form-field-validator';
-export * from './dynamic-table.model';
-export * from './dynamic-table-column';
-export * from './dynamic-table-row';
