+++ ng2-components/ng2-activiti-form/src/components/widgets/display-value/display-value.widget.spec.ts
@@ -21,8 +21,7 @@ import { FormService } from '../../../services/form.service';
 import { FormFieldModel } from './../core/form-field.model';
 import { FormFieldTypes } from '../core/form-field-types';
 import { FormModel } from '../core/form.model';
-import { DynamicTableRow } from './../core/dynamic-table-row';
-import { DynamicTableColumn } from './../core/dynamic-table-column';
+import { DynamicTableColumn, DynamicTableRow } from './../dynamic-table/dynamic-table.widget.model';
 
 describe('DisplayValueWidget', () => {
