+++ ng2-components/ng2-activiti-form/src/components/widgets/dynamic-table/editors/text/text.editor.spec.ts
@@ -16,7 +16,7 @@
  */
 
 import { TextEditorComponent } from './text.editor';
-import { DynamicTableRow, DynamicTableColumn } from './../../../core/index';
+import { DynamicTableRow, DynamicTableColumn } from './../../dynamic-table.widget.model';
 
 describe('TextEditorComponent', () => {
